Moved content to single .cpp
